[ICU-12570] remove undefined behaviors from cintltst. Mostly signed integer overflow. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39042 Former-commit-id: 056ddf8fbb1f09e6ad9f2194eb5ff02c3cf05ff8

+++ icu4c/source/test/cintltst/bocu1tst.c
@@ -344,7 +344,7 @@ packDiff(int32_t diff) {
     } while(--count>0);
 
     /* add lead byte */
-    result|=(lead+diff)<<shift;
+    result |= (uint32_t)(lead+diff)<<shift;
 
     return result;
 }
