+++ icu4c/source/test/cintltst/sorttest.c
@@ -53,7 +53,7 @@ SortTest() {
 
     /* for medium, add bits that will not be compared, to test stability */
     for(i=0; i<UPRV_LENGTHOF(medium); ++i) {
-        medium[i]=(medium[i]<<4)|i;
+        medium[i]=(int32_t)((uint32_t)medium[i]<<4) | i;
     }
 
     /* sort medium array (stable) */
